backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.8.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
PyJWT>=2.10.0,<3.0.0
python-multipart>=0.0.20,<0.1.0

# LLM & RAG
langchain>=0.1.0,<1.0.0
langchain-core>=1.2.0,<2.0.0
langchain-groq>=1.1.0,<2.0.0
langchain-community>=0.0.10,<1.0.0
groq>=0.20.0,<1.0.0

# Vector Search & Embeddings
faiss-cpu>=1.7.4,<2.0.0
sentence-transformers>=3.0.0,<4.0.0
numpy>=1.26.0,<2.0.0

# Memory & Cache
redis>=5.0.0,<6.0.0

# Security
bcrypt>=4.2.0,<5.0.0
passlib>=1.7.0,<2.0.0

# Validation
email-validator>=2.2.0,<3.0.0

# Utilities
requests>=2.32.0,<3.0.0
httpx>=0.28.0,<1.0.0
python-dateutil>=2.8.0,<3.0.0
